+++ src/main/java/com/jslog_spring/domain/post/service/PostServiceImpl.java
@@ -13,20 +13,20 @@ import java.util.Optional;
 public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
 
-    public Post createPost(int authorId, String title, String content) {
+    public Post createPost(Long authorId, String title, String content) {
         Post newPost = Post.create(authorId, title, content);
         return postRepository.save(newPost);
     }
 
-    public Post updatePost(int postId, String title, String content) {
+    public Post updatePost(Long postId, String title, String content) {
         return null;
     }
 
-    public void deletePost(int postId) {
+    public void deletePost(Long postId) {
 
     }
 
-    public Optional<Post> getPost(int postId) {
+    public Optional<Post> getPost(Long postId) {
         return Optional.empty();
     }
 
@@ -34,7 +34,7 @@ public class PostServiceImpl implements PostService {
         return List.of();
     }
 
-    public List<Post> getPostsByAuthor(int authorId, int page, int size) {
+    public List<Post> getPostsByAuthor(Long authorId, int page, int size) {
         return List.of();
     }
 
@@ -42,15 +42,17 @@ public class PostServiceImpl implements PostService {
         return List.of();
     }
 
-    public List<Post> searchPostsByAuthor(int authorId, String keyword, int page, int size) {
+    public List<Post> searchPostsByAuthor(Long authorId, String keyword, int page, int size) {
         return List.of();
     }
 
-    public long countPostsByAuthor(int authorId) {
+    public long countPostsByAuthor(Long authorId) {
         return 0;
     }
 
     public long countAllPosts() {
         return 0;
     }
+
+
 }
